Move JS files to src subfolder
